[MPOM-270] Fix enforcer plugin (#32)
Enforcer as is causes that downstream in case of update
POM must completely redefine plugin, thus, loosing all
the benefits of parent POM pluginMgmt.

Move plugin definition to pluginMgmt instead to fully
define it and it's dependents in plugin section. Keep
execution in plugin section.

Update to latest enforcer as well.