[MPOM-374] Apply spotless to groovy
This extends spotless configuration to also check groovy files.
As our groovy files for IT are not in the default groovy packages include is overwritten.
The delimiter is overwritten, because the empty package is not supported by spotbugs, even while it's a valid package.

see also: https://github.com/diffplug/spotless/issues/1418

closes #284